Force secret by default
Normally secrets are passed via environment variables, but some odd case with GitHub actions and then docker seems to not be finding the environment variable. 

The variable is only use for redirection of the authentication, but again this has to work for production which is does if manually defined and ran with docker locally.